# contracts/openapi/v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenLibraryOS API",
    "version": "1.0.0",
    "description": "The versioned public API contract. Breaking changes require a new API major version."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Authenticate in the named organization.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessToken"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationFailed"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refreshAccessToken",
        "summary": "Rotate a refresh session and issue a new access token.",
        "security": [
          {
            "CsrfToken": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessToken"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationFailed"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Revoke the current refresh session chain.",
        "security": [
          {
            "BearerAuth": [],
            "CsrfToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "The session was revoked.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/X-Request-ID"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationFailed"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "getCurrentPrincipal",
        "summary": "Return only the authenticated user, organization, and session identifiers.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authenticated principal and live effective permissions.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/X-Request-ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessPrincipal"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationFailed"
          }
        }
      }
    },
...
    "/auth/password/change": {
      "post": {
        "operationId": "changePassword",
        "summary": "Change the authenticated user's password.",
        "security": [
          {
            "BearerAuth": [],
            "CsrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The password was changed.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/X-Request-ID"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationFailed"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "X-Request-ID": {
        "description": "Server-generated correlation identifier, or a syntactically valid client value.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "request_id"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "request_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._-]{1,128}$"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "organization_slug",
          "email",
          "password"
        ],
        "properties": {
          "organization_slug": {
            "type": "string",
            "pattern": "^[a-z0-9](?:[a-z0-9-]{0,62}[a-z0-9])?$"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "additionalProperties": false
      },
...
      "AccessToken": {
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "const": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1
          }
        },
        "additionalProperties": false
      },
...
      "AccessPrincipal": {
        "type": "object",
        "required": [
          "user_id",
          "organization_id",
          "session_id",
          "permissions"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "session_id": {
            "type": "string",
            "format": "uuid"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "additionalProperties": false
      },
      "PasswordChangeRequest": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string",
            "format": "password"
          },
          "new_password": {
            "type": "string",
            "format": "password",
            "minLength": 12
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
...
      "AccessToken": {
        "description": "A short-lived access token.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/X-Request-ID"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AccessToken"
            }
          }
        }
      },
      "AuthenticationFailed": {
        "description": "Authentication failed without revealing account or tenant state.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/X-Request-ID"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "ProblemDetails": {
        "description": "An RFC Problem Details error response.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/X-Request-ID"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  }
}